backend/requirements.txt
django
gunicorn
whitenoise
psycopg2-binary
requests
urllib3
python-dotenv
psycopg[binary]
dj-database-url
djangorestframework
djangorestframework-simplejwt
django-filter
django-cors-headers